added a cast

+++ src/org/escaperun/game/model/entities/statistics/OffensiveRating.java
@@ -20,7 +20,7 @@ public class OffensiveRating extends DerivedStatistic<Integer>{
     protected void recalculate() {
         //idk I made something up
         //also need a way to notify when weapon changes
-        offensiveRating = (3* str.getCurrent() + 2*lvl.getCurrent()) + weaponDamage;
+        offensiveRating = (double) (3* str.getCurrent() + 2*lvl.getCurrent()) + weaponDamage;
     }
